# Conflicts:
#	TIGLViewer/src/TIGLViewerContext.cpp
#	TIGLViewer/src/TIGLViewerContext.h
#	TIGLViewer/src/TIGLViewerDocument.cpp
#	TIGLViewer/src/TIGLViewerDocument.h
#	TIGLViewer/src/TIGLViewerWindow.cpp
#	src/api/tigl.cpp
#	src/api/tigl.h
#	src/common/tiglcommonfunctions.cpp
#	src/common/tiglcommonfunctions.h
#	src/exports/CTiglExportVtk.cpp
#	src/wing/CCPACSWing.cpp
#	src/wing/CCPACSWing.h
#	src/wing/CCPACSWingComponentSegment.cpp
#	src/wing/CCPACSWingComponentSegment.h
#	src/wing/CCPACSWingSegment.cpp
#	src/wing/CCPACSWingSegment.h
#	tests/tiglCommonFunctions.cpp
#	tests/tiglPointTranslator.cpp

+++ src/cpacs_other/CCPACSExternalObjects.cpp
@@ -84,7 +84,7 @@ CCPACSExternalObject&CCPACSExternalObjects::GetObject(int index) const
 
 int CCPACSExternalObjects::GetObjectCount() const
 {
-    return _objects.size();
+    return static_cast<int>(_objects.size());
 }
 
     // Cleanup routine
